+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 * Renamed `SwiftTemplate` to `StencilSwiftTemplate`.  
   [David Jennes](https://github.com/djbe)
   [#14](https://github.com/SwiftGen/StencilSwiftKit/issues/14)
+* Refactor stencil swift extensions registration for easier use with an existing `Extension`.  
+  [David Jennes](https://github.com/djbe)
+  [#16](https://github.com/SwiftGen/StencilSwiftKit/issues/16)
 
 ## Before 5.0.0